Migrate more things to their own components
